started to refactor main procedure

+++ Source/Character.cpp
@@ -13,7 +13,7 @@ Character::~Character()
 {
 }
 
-double Character::GetMovespeed()
+int Character::GetMovespeed()
 {
 	return movespeed;
 }
